# backend/docs/v3/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : { },
    "description" : "This is the API documentation for the Muzee application.",
    "title" : "Muzee API",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost/api"
  } ],
  "paths" : {
...
    "/v1/auth/login" : {
      "post" : {
        "description" : "Authenticates user with email and password. Returns tokens in JSON for mobile clients, and sets HttpOnly cookies for web browsers. Requires client_id for session tracking.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/internal_interface_handler.LoginRequest"
              }
            }
          },
          "description" : "Email, password, and client ID",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/internal_interface_handler.LoginResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Too Many Requests"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "User login",
        "tags" : [ "auth" ],
        "x-codegen-request-body-name" : "request"
      }
    },
    "/v1/auth/logout" : {
      "post" : {
        "description" : "Invalidates the refresh token and ends the user's session. Accepts refresh token from either HttpOnly cookie (web) or request body (mobile). Also clears cookies for web browsers.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/internal_interface_handler.LogoutRequest"
              }
            }
          },
          "description" : "Refresh token (optional if using cookies)",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/internal_interface_handler.LogoutResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "User logout",
        "tags" : [ "auth" ],
        "x-codegen-request-body-name" : "request"
      }
    },
    "/v1/auth/refresh" : {
      "post" : {
        "description" : "Generates new access and refresh tokens. Accepts refresh token from either HttpOnly cookie (web) or request body (mobile). The old refresh token is invalidated. Requires client_id for session validation.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/internal_interface_handler.RefreshTokenRequest"
              }
            }
          },
          "description" : "Refresh token and client ID (refresh_token optional if using cookies)",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/internal_interface_handler.RefreshTokenResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Refresh access token",
        "tags" : [ "auth" ],
        "x-codegen-request-body-name" : "request"
      }
    },
    "/v1/auth/signup/resend-code" : {
      "post" : {
        "description" : "Resends a new 6-digit verification code to the email for signup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/internal_interface_handler.ResendCodeRequest"
              }
            }
          },
          "description" : "Email",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/internal_interface_handler.ResendCodeResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Too Many Requests"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Resend verification code",
        "tags" : [ "auth" ],
        "x-codegen-request-body-name" : "request"
      }
    },
    "/v1/auth/signup/send-code" : {
      "post" : {
        "description" : "Sends a 6-digit verification code to the email for signup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/internal_interface_handler.SendCodeRequest"
              }
            }
          },
          "description" : "Email and password",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/internal_interface_handler.SendCodeResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Too Many Requests"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_keu-5_muzee_backend_internal_helper.ErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Send verification code",
        "tags" : [ "auth" ],
        "x-codegen-request-body-name" : "request"
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "github_com_keu-5_muzee_backend_internal_helper.ErrorResponse" : {
        "properties" : {
          "details" : {
            "items" : {
              "additionalProperties" : true,
              "type" : "object"
            },
            "type" : "array"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "internal_interface_handler.LoginRequest" : {
        "properties" : {
          "client_id" : {
            "maxLength" : 255,
            "minLength" : 1,
            "type" : "string"
          },
          "email" : {
            "maxLength" : 255,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          }
        },
        "required" : [ "client_id", "email", "password" ],
        "type" : "object"
      },
      "internal_interface_handler.LoginResponse" : {
        "properties" : {
          "access_token" : {
            "type" : "string"
          },
          "expires_in" : {
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          },
          "refresh_token" : {
            "type" : "string"
          },
          "token_type" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/internal_interface_handler.UserResponse"
          }
        },
        "type" : "object"
      },
      "internal_interface_handler.LogoutRequest" : {
        "properties" : {
          "refresh_token" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "internal_interface_handler.LogoutResponse" : {
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "internal_interface_handler.RefreshTokenRequest" : {
        "properties" : {
          "client_id" : {
            "maxLength" : 255,
            "minLength" : 1,
            "type" : "string"
          },
          "refresh_token" : {
            "type" : "string"
          }
        },
        "required" : [ "client_id" ],
        "type" : "object"
      },
      "internal_interface_handler.RefreshTokenResponse" : {
        "properties" : {
          "access_token" : {
            "type" : "string"
          },
          "expires_in" : {
            "type" : "integer"
          },
          "refresh_token" : {
            "type" : "string"
          },
          "token_type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "internal_interface_handler.ResendCodeRequest" : {
        "properties" : {
          "email" : {
            "maxLength" : 255,
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  },
...
}